# internal/spec/openapi.yaml
openapi: "3.0.3"
info:
  title: Identity API
  version: "1.0.0"
  description: |
    Self-hosted JWT authentication and identity management service.

    ## Token model
    - **Access token**: short-lived JWT (15 minutes). Pass in `Authorization: Bearer <token>`. Never persist to disk.
    - **Refresh token**: long-lived opaque token (30-day sliding window). Rotated on every use — always persist the newest one immediately.

    ## Token family rotation & theft detection
    Every refresh call invalidates the submitted token and issues a new pair. If a previously-used (rotated) refresh token is ever presented again, the server treats it as a stolen token: the **entire token family** is revoked and `token_family_compromised` is returned. The legitimate client must log in again.

    ## Refresh implementation note
    Use a mutex around the refresh call. If multiple in-flight requests all receive a 401 simultaneously, only one should call `/auth/refresh`; the rest should wait and reuse the result. See docs/api.md for pseudocode.
servers:
  - url: https://{domain}/api/v1
    variables:
      domain:
        default: id.swee.net
        description: Your deployment hostname
security: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in, refresh_token]
      properties:
        access_token:
          type: string
          description: Short-lived JWT. Use in Authorization header. Do NOT persist to disk.
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Seconds until access token expires. Always 900 (15 minutes).
          example: 900
        refresh_token:
          type: string
          description: Long-lived opaque token. Store securely (Keychain / EncryptedSharedPreferences). Rotated on every use.
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: |
            Snake_case error code. Always check this field, not the HTTP status alone.
            Known codes: invalid_credentials, account_disabled, invalid_refresh_token,
            token_expired, token_family_compromised, forbidden, not_found, username_taken,
            cannot_delete_last_admin, validation_error, user_limit_reached, weak_password,
            rate_limited, internal_error, webauthn_not_enabled, webauthn_invalid_challenge,
            webauthn_verification_failed, webauthn_credential_not_found, webauthn_credential_limit
        message:
          type: string
          description: Human-readable description. Do not parse programmatically.
...
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Rotate token pair
      description: |
        Exchange a refresh token for a new access token and refresh token.
        The submitted refresh token is **immediately invalidated**.
        Replaying a used token triggers `token_family_compromised` and revokes all sessions in that family.
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Tokens rotated — persist the new refresh token immediately
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "401":
          description: Token invalid, expired, or family compromised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalid_refresh_token:
                  value:
                    error: invalid_refresh_token
                    message: the refresh token is invalid
                token_expired:
                  value:
                    error: token_expired
                    message: the refresh token has expired
                token_family_compromised:
                  value:
                    error: token_family_compromised
                    message: "token reuse detected — please log in again"
        "403":
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /oauth/authorize:
    get:
      operationId: oauthAuthorize
      summary: Render OAuth login form
      description: |
        Validates the client and redirect URI, then renders a login form.
        This endpoint is at `/oauth/authorize` (not under `/api/v1`).
        Only `code_challenge_method=S256` is accepted.
      tags: [oauth]
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Login form rendered (HTML)
        "400":
          description: Invalid client, redirect URI, or missing parameters (HTML error page)

    post:
      operationId: oauthAuthorizeSubmit
      summary: Submit OAuth login credentials
      description: |
        Authenticates the user and redirects to the client's redirect URI with an authorization code.
        The code is single-use and expires after 60 seconds.
      tags: [oauth]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [client_id, redirect_uri, code_challenge, username, password]
              properties:
                client_id:
                  type: string
                redirect_uri:
                  type: string
                state:
                  type: string
                code_challenge:
                  type: string
                username:
                  type: string
                password:
                  type: string
      responses:
        "302":
          description: Redirect to client with `?code=...&state=...`
        "200":
          description: Login form re-rendered with error (bad credentials)

  # ── WebAuthn / Passkey endpoints ───────────────────────────────
...
  /oauth/token:
    post:
      operationId: oauthToken
      tags: [oauth]
      summary: Token endpoint
      description: |
        Supports three grant types:
        - `authorization_code` — exchange an auth code + PKCE verifier for tokens
        - `refresh_token` — rotate a refresh token for a new token pair
        - `client_credentials` — service authenticates with client_id + client_secret, receives an access token (no refresh token)

        For `client_credentials`, authenticate via `Authorization: Basic base64(client_id:client_secret)` or form body.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type]
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code, refresh_token, client_credentials]
                scope:
                  type: string
                  description: Space-delimited scopes (client_credentials only)
                client_id:
                  type: string
                client_secret:
                  type: string
                code:
                  type: string
                redirect_uri:
                  type: string
                code_verifier:
                  type: string
                refresh_token:
                  type: string
      responses:
        "200":
          description: Token(s) issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    example: Bearer
                  expires_in:
                    type: integer
                    example: 900
                  refresh_token:
                    type: string
                  scope:
                    type: string
        "400":
          description: Invalid request, unsupported grant type, invalid scope, or unauthorized client
        "401":
          description: Invalid client credentials
  /oauth/introspect:
    post:
      tags: [oauth]
      summary: Token introspection (RFC 7662)
      description: Validates a token and returns its claims. Requires client authentication.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Introspection result
          content:
            application/json:
              schema:
                type: object
                required: [active]
                properties:
                  active:
                    type: boolean
                  sub:
                    type: string
                  client_id:
                    type: string
                  scope:
                    type: string
                  token_type:
                    type: string
        "401":
          description: Missing or invalid client credentials
...